Moving SpecialAbilityGenerator
