+++ MekHQ/src/mekhq/campaign/personnel/generator/DefaultSpecialAbilityGenerator.java
@@ -16,10 +16,12 @@
  * You should have received a copy of the GNU General Public License
  * along with MekHQ.  If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.campaign.personnel;
+package mekhq.campaign.personnel.generator;
 
 import mekhq.Utilities;
 import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.personnel.Person;
+import mekhq.campaign.personnel.SkillType;
 
 public class DefaultSpecialAbilityGenerator extends AbstractSpecialAbilityGenerator {
